airflow/requirements.txt
# === Airflow Core & Providers ===
apache-airflow==2.8.1
apache-airflow-providers-postgres>=5.0.0
apache-airflow-providers-http
airflow-provider-mlflow==1.1.0

# === External API Clients ===
requests
openai
praw
yfinance

# === JSON Parsing & Date Handling ===
python-dateutil
simplejson

# === Database Driver ===
psycopg2-binary

# === Env File Support ===
python-dotenv

# === NLP & Machine Learning ===
transformers
torch
langdetect
scikit-learn
bertopic
hf_xet
tiktoken>=0.5.1
nltk

# === Optional (for sentence-transformers used in BERTopic) ===
sentence-transformers